test: add unit tests
